+++ CHANGELOG.md
@@ -7,9 +7,11 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Include chart name when installing with a custom release name ([#233], [#234]).
+- `operator-rs` `0.21.1` -> `0.22.0` ([#235]).
 
 [#233]: https://github.com/stackabletech/trino-operator/pull/233
 [#234]: https://github.com/stackabletech/trino-operator/pull/234
+[#235]: https://github.com/stackabletech/trino-operator/pull/235
 
 ## [0.4.0] - 2022-06-30